% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseClusterRes.R
\name{chooseClusterRes}
\alias{chooseClusterRes}
\title{Allows experimentation of different cluster resolutions on a Seurat object.}
\usage{
chooseClusterRes(
  seurat.obj,
  cluster.res,
  ident.plot = TRUE,
  feature.plot = NULL,
  plot.cols = NULL
)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{cluster.res}{Cluster resolutions interested in}

\item{feature.plot}{Produce a feature plot based on given genes}
}
\value{
UMAP and a Feature Plot
}
\description{
Allows experimentation of different cluster resolutions on a Seurat object.
}
\examples{
chooseClusterRes(seurat.obj, cluster.res = list(.1,.2,.5), feature.plot = c("gene1","gene2","gene3"))
}
