% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crToMTX.R
\name{crToMTX}
\alias{crToMTX}
\title{Takes a CellRanger filtered output and returns a gene matrix}
\usage{
crToMTX(directory)
}
\arguments{
\item{directory}{The path to the directory that contains the filtered_gene_matrices folder}
}
\value{
A filtered gene matrix
}
\description{
Takes a CellRanger filtered output and returns a gene matrix
}
\examples{
directory <- "/home/users/documents/CRoutput"
matrix <- crToMTX(directory) 
}
