% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crToSeurat.R
\name{crToSeurat}
\alias{crToSeurat}
\title{Takes a directory CellRanger counts output and returns a list of Seurat objects or a merged Seurat object}
\usage{
crToSeurat(
  directory,
  min.cells = 3,
  min.features = 200,
  sample.names = c("sample"),
  merge = FALSE
)
}
\arguments{
\item{seurat.obj}{A Seurat object.}

\item{min_cells}{Include features detected in at least this many cells.}

\item{min_features}{Include cells where at least this many features are detected}
}
\value{
A Seurat object
}
\description{
Takes a directory CellRanger counts output and returns a list of Seurat objects or a merged Seurat object
}
\examples{
crToSeurat(directory = "Files", min_cells = 3, min_features = 3)
}
