% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendoSeurat.R
\name{dendoSeurat}
\alias{dendoSeurat}
\title{Produce hierarchical clustering for a sub-cluster of a downsampled Seurat object and return a dendrogram.}
\usage{
dendoSeurat(
  seurat.obj,
  cluster,
  annotation.name,
  down.sample,
  variable.genes = NULL,
  seed = 1,
  return.clusters = NULL
)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{cluster}{Cluster interested in}

\item{annotation.name}{Variable name given to Seurat cluster assignments}

\item{down.sample}{Amount of cells to sample}

\item{variable.genes}{(Optional) Subet counts data to this many variable genes for distance matrix calculation}

\item{seed}{Value for the seed to set}

\item{return.clusters}{(Optional) Put the height you want to cut the dendrogram at and return an object that contains cells and hierarchical clusters
dendoSeurat(seurat.obj, cluster = "5", annotation.name = "seurat_clusters", down.sample = 50, seed = 1, return.clusters = 4)}
}
\description{
Produce hierarchical clustering for a sub-cluster of a downsampled Seurat object and return a dendrogram.
}
