% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaExtraction.R
\name{extractCounts}
\alias{extractCounts}
\title{Easily extract counts from a Seurat object}
\usage{
extractCounts(seurat.obj, assay = "SCT", genes = NULL, tibble = FALSE)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{assay}{The assay to pull from}

\item{genes}{Optional argument to extract certain genes only}

\item{tibble}{Whether you want counts returned as a tibble}
}
\value{
Counts
}
\description{
Easily extract counts from a Seurat object
}
\examples{
extractCounts(seurat.obj, assay = "RNA", genes = c("TBC1D3D","LINC00470"), tibble = TRUE) 
}
