% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaExtraction.R
\name{extractMeta}
\alias{extractMeta}
\title{Easily extract Seurat meta-data into a tibble}
\usage{
extractMeta(seurat.obj, types, vars, merge = TRUE)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{types}{The types of meta-data you want to extract. Options are c("Embeddings","Clusters","Module Scores")}

\item{vars}{The names within the Seurat object of the meta-data you want to extract.}

\item{merge}{Whether you want all data to merged. If set to TRUE, data will be merged by Cells.}
}
\value{
Meta-data in tibble format
}
\description{
Easily extract Seurat meta-data into a tibble
}
