% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierachialSeurat.R
\name{hierachialSeurat}
\alias{hierachialSeurat}
\title{Produce hierachial clustering for a sub-cluster of Seurat object. Due to the sheer computational demands of hierachial clustering with single cell data, 
Seurat object must be downsampled.}
\usage{
hierachialSeurat(
  seurat.obj,
  clusters,
  annotation.name,
  down.sample,
  variable.genes = NULL
)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{clusters}{Cluster to cluster}

\item{annotation.name}{Variable name given to Seurat cluster assignments}

\item{down.sample}{Amount of cells to sample}

\item{variable.genes}{(Optional) Subet counts data to this many variable genes for distance matrix calculation
hierachialSeurat(seurat.obj, clusters = "Cluster1", annotation.name = "Seurat_Assignment", down.sample = 50)}
}
\description{
Produce hierachial clustering for a sub-cluster of Seurat object. Due to the sheer computational demands of hierachial clustering with single cell data, 
Seurat object must be downsampled.
}
