% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSubCluster.R
\name{mergeSubCluster}
\alias{mergeSubCluster}
\title{Merge a reclusted sub-sample back into main object}
\usage{
mergeSubCluster(seurat.obj, subcluster.obj, cluster.replace, annotation.name)
}
\arguments{
\item{seurat.obj}{A Seurat object.}

\item{subcluster.obj}{A Seurat sub-cluster object}

\item{annotation.name}{Name to store cluster annotations}

\item{cluster.repace}{Clusters to replace}
}
\value{
Merged seurat object
}
\description{
Merge a reclusted sub-sample back into main object
}
\examples{
mergedSubCluster(seurat.obj, subcluster.obj = seurat.sub.obj, cluster.replace = "T Cells", annotation.name = "Seurat_Assignment") 
}
