% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSeurat.R
\name{processSeurat}
\alias{processSeurat}
\title{Pipes together several downstream Seurat steps including variance-stabilizing transformation, PCA, clustering and nonlinear dimensionality reduction}
\usage{
processSeurat(
  seurat.obj,
  dims = 1:50,
  cluster.res = 0.2,
  vars_to_regress = c("orig.ident", "percent.mt"),
  seed.use = 24,
  n.cores = NULL
)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{dims}{Dimensions of reduction to use as input for FindNeighbors and RunUMAP}

\item{cluster.res}{Cluster resolution for FindClusters}

\item{vars_to_regress}{Variables to regress out in a second non-regularized linear regression for SCTransform}

\item{seed.use}{Seed to set for RunUMAP}

\item{n.cores}{Number of cores to use}
}
\value{
A processed Seurat object
}
\description{
Pipes together several downstream Seurat steps including variance-stabilizing transformation, PCA, clustering and nonlinear dimensionality reduction
}
\examples{
processSeurat(seurat.obj, dims = 1:50, cluster.res = .2, vars_to_regress = c("orig.ident","percent.mt"))
}
