% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameClusters.R
\name{renameClusters}
\alias{renameClusters}
\title{Easily rename/annotate Seurat clusters}
\usage{
renameClusters(seurat.obj, cluster.names, annotation.name = NULL)
}
\arguments{
\item{seurat.obj}{A Seurat object.}

\item{cluster.names}{Names for the renamed/annotated clusters}

\item{annotation.name}{Name to store new clusters in Seurat object}
}
\value{
A Seurat object
}
\description{
Easily rename/annotate Seurat clusters
}
\examples{
renameClusters(seurat.obj, cluster.names = c("Cluster1","Cluster2","Cluster3"), annotation.name = "Seurat_Annotations")
}
