% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{umapCellAnno}
\alias{umapCellAnno}
\title{Produces a UMAP plot from a Seurat object that can put cluster cell amounts in the legend, title, or on the plot.}
\usage{
umapCellAnno(
  seurat.obj,
  annotation.name = NULL,
  point.size = 1,
  label.size = 10,
  title = "",
  legend.title.size = 0,
  legend.text.size = 15,
  axis.title.x.size = 15,
  axis.title.y.size = 15,
  axis.text.y.left.size = 15,
  axis.text.x.bottom.size = 15,
  counts.as.title = FALSE,
  legend = TRUE,
  cell.legend.size = 10,
  counts.in.legend = TRUE,
  use.cols = NULL
)
}
\arguments{
\item{annotation.name}{The name of the slot that contains the Seurat clusters. Defaults to "Seurat_Assignment" first, then "seurat_clusters"}

\item{point.size}{Size of the UMAP plot points}

\item{title}{Title of the plot}

\item{legend.title.size}{Size of the legend title}

\item{legend.text.size}{Size of the legend text}

\item{axis.title.x.size}{X-axis title text size}

\item{axis.title.y.size}{Y-axis title text size}

\item{axis.text.x.bottom.size}{Y-axis sub-text size}

\item{counts.as.title}{Put the total amount of cells as the title}

\item{legend}{Show the legend}

\item{cell.legend.size}{Show the legend}

\item{counts.in.legend}{Put the amount of cells alongside cluster name in the legend}

\item{use.cols}{Use custom colors for the cell clusters}

\item{axis.text.x.left.size}{X-axis sub-text size}
}
\value{
A UMAP plot
}
\description{
Produces a UMAP plot from a Seurat object that can put cluster cell amounts in the legend, title, or on the plot.
}
\examples{
umapCellAnno(seurat.obj, annotation.name = "Seurat_Clusters")
}
