% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{variableGeneMatrix}
\alias{variableGeneMatrix}
\title{Produces a variable gene correlation heatmap and returns a correlation matrix.}
\usage{
variableGeneMatrix(
  seurat.obj,
  variable.genes = 200,
  return.table = FALSE,
  fontsize = 10
)
}
\arguments{
\item{seurat.obj}{A seurat object.}

\item{variable.genes}{The amount of variables genes to include.}

\item{return.table}{Optional argument to return the correlation matrix}

\item{fontsize}{The font size for the gene names on the heatmap.}

\item{downsample}{Amount of cells to optionally downsample.}
}
\value{
A correlation heatmap and (optionally) a correlation matrix.
}
\description{
Produces a variable gene correlation heatmap and returns a correlation matrix.
}
\examples{
variableGeneMatrix(seurat.obj = seurat.obj, variable.genes = 200, return.table = FALSE, fontsize = 10) 
}
